This should be a stash, except i dont know the syntax and dont have the time to learn now.
i did a lot of formatting changes, so lots of lines have changed
and then i went and added more operators
and then i added some syntactic sugar to them.

this last part is not yet working.

the sample wordpic has been enhanced to reflect all these possible relation expressions, and one of them is failing
to be continued